requirements.txt
# CS2 Arbitrage Bot Dependencies
# Install with: pip install -r requirements.txt

# Database ORM
sqlalchemy>=2.0.0

# HTTP client
httpx>=0.25.0

# CLI framework
typer>=0.9.0

# Rich terminal UI
rich>=13.0.0

# Environment configuration
python-dotenv>=1.0.0

# Data validation
pydantic>=2.0.0

# YAML config parsing
pyyaml>=6.0.0

# Development dependencies (optional)
# pytest>=7.0.0
# black>=23.0.0
# ruff>=0.1.0
# mypy>=1.0.0

